remove unnecessary typed aliases
The compiler now imports these as their canonical type, allowing us to
directly use without the explicit type cast.

Introduce two explicit invalid casts as the shape of the current API
does not account for the error code type difference (Windows uses an
unsigned error code).

+++ Sources/PackageRegistryCommand/PackageRegistryCommand+Auth.swift
@@ -44,7 +44,7 @@ private func readpassword(_ prompt: String) throws -> String {
 
     print(prompt, terminator: "")
 
-    guard SetConsoleMode(hStdIn, DWORD(ENABLE_LINE_INPUT)) else {
+    guard SetConsoleMode(hStdIn, ENABLE_LINE_INPUT) else {
         throw StringError("unable to read input: SetConsoleMode failed")
     }
     defer { SetConsoleMode(hStdIn, dwMode) }
